requirements.txt
openai-whisper>=20231117
sentence-transformers>=2.2.2
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0
safetensors>=0.4.0
matplotlib>=3.7.0
ffmpeg-python>=0.2.0
pillow>=10.0.0
tqdm>=4.66.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
notion-client>=2.0.0
python-dotenv>=1.0.0
rich>=13.4.2
typer>=0.9.0
